M: nodejs 端 model update 接口支持更新模型文件 && 前端目前还不支持修改模型文件

+++ server/utils/index.js
@@ -5,6 +5,8 @@ const delay = (duration = 1000) =>
     new Promise(resolve => setTimeout(resolve, duration))
 
 const rmdirSync = dir => {
+    if (!fs.existsSync(dir)) return
+
     const fileinfo = fs.statSync(dir)
     if (fileinfo.isFile()) {
         fs.unlinkSync(dir)
